Move Views from views/ to routes/
