+++ src/routes/ViewPlaylistDetails.tsx
